Moved PhpExtractor and PhpStringTokenParser to Translation component
