+++ src/Symfony/Bundle/FrameworkBundle/Translation/PhpStringTokenParser.php
@@ -11,132 +11,11 @@
 
 namespace Symfony\Bundle\FrameworkBundle\Translation;
 
-/*
- * The following is derived from code at http://github.com/nikic/PHP-Parser
- *
- * Copyright (c) 2011 by Nikita Popov
- *
- * Some rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are
- * met:
- *
- *     * Redistributions of source code must retain the above copyright
- *       notice, this list of conditions and the following disclaimer.
- *
- *     * Redistributions in binary form must reproduce the above
- *       copyright notice, this list of conditions and the following
- *       disclaimer in the documentation and/or other materials provided
- *       with the distribution.
- *
- *     * The names of the contributors may not be used to endorse or
- *       promote products derived from this software without specific
- *       prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+@trigger_error(sprintf('The class "%s" is deprecated since version 3.4 and will be removed in 4.0. Use "%s" instead. ', PhpStringTokenParser::class, \Symfony\Component\Translation\Extractor\PhpStringTokenParser::class), E_USER_DEPRECATED);
+
+/**
+ * @deprecated since version 3.4 and will be removed in 4.0. Use Symfony\Component\Translation\Extractor\PhpStringTokenParser instead
  */
-
-class PhpStringTokenParser
+class PhpStringTokenParser extends \Symfony\Component\Translation\Extractor\PhpStringTokenParser
 {
-    protected static $replacements = array(
-        '\\' => '\\',
-        '$' => '$',
-        'n' => "\n",
-        'r' => "\r",
-        't' => "\t",
-        'f' => "\f",
-        'v' => "\v",
-        'e' => "\x1B",
-    );
-
-    /**
-     * Parses a string token.
-     *
-     * @param string $str String token content
-     *
-     * @return string The parsed string
-     */
-    public static function parse($str)
-    {
-        $bLength = 0;
-        if ('b' === $str[0]) {
-            $bLength = 1;
-        }
-
-        if ('\'' === $str[$bLength]) {
-            return str_replace(
-                array('\\\\', '\\\''),
-                array('\\', '\''),
-                substr($str, $bLength + 1, -1)
-            );
-        } else {
-            return self::parseEscapeSequences(substr($str, $bLength + 1, -1), '"');
-        }
-    }
-
-    /**
-     * Parses escape sequences in strings (all string types apart from single quoted).
-     *
-     * @param string      $str   String without quotes
-     * @param null|string $quote Quote type
-     *
-     * @return string String with escape sequences parsed
-     */
-    public static function parseEscapeSequences($str, $quote)
-    {
-        if (null !== $quote) {
-            $str = str_replace('\\'.$quote, $quote, $str);
-        }
-
-        return preg_replace_callback(
-            '~\\\\([\\\\$nrtfve]|[xX][0-9a-fA-F]{1,2}|[0-7]{1,3})~',
-            array(__CLASS__, 'parseCallback'),
-            $str
-        );
-    }
-
-    private static function parseCallback($matches)
-    {
-        $str = $matches[1];
-
-        if (isset(self::$replacements[$str])) {
-            return self::$replacements[$str];
-        } elseif ('x' === $str[0] || 'X' === $str[0]) {
-            return chr(hexdec($str));
-        } else {
-            return chr(octdec($str));
-        }
-    }
-
-    /**
-     * Parses a constant doc string.
-     *
-     * @param string $startToken Doc string start token content (<<<SMTHG)
-     * @param string $str        String token content
-     *
-     * @return string Parsed string
-     */
-    public static function parseDocString($startToken, $str)
-    {
-        // strip last newline (thanks tokenizer for sticking it into the string!)
-        $str = preg_replace('~(\r\n|\n|\r)$~', '', $str);
-
-        // nowdoc string
-        if (false !== strpos($startToken, '\'')) {
-            return $str;
-        }
-
-        return self::parseEscapeSequences($str, null);
-    }
 }
