+++ src/Symfony/Component/Translation/CHANGELOG.md
@@ -13,6 +13,8 @@ CHANGELOG
  * Added `TranslationWriterInterface`
  * Deprecated `TranslationWriter::writeTranslations` in favor of `TranslationWriter::write`
  * added support for adding custom message formatter and decoupling the default one.
+ * Added `PhpExtractor`
+ * Added `PhpStringTokenParser`
 
 3.2.0
 -----
